Move eventLogs and ens utils to ~utils/web3
